pyproject.toml
[project]
name = "youtube-videos-recommender"
version = "0.1.0"
description = "A YT videoes recommender powered by LLMs."
authors = [
    {name = "Skander Kort <https://skanderkort.com>"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client (>=2.170.0,<3.0.0)",
    "google-auth (>=2.40.2,<3.0.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "groq (>=1.0.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
